Emitter Tests is just separated to run SapiEmitter specific test - BufferLevel test
