+++ test/Response/SapiEmitterTest.php
@@ -9,71 +9,30 @@
 
 namespace ZendTest\Diactoros\Response;
 
-use PHPUnit_Framework_TestCase as TestCase;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\StreamInterface;
 use Zend\Diactoros\Response;
-use Zend\Diactoros\Response\SapiEmitter;
-use ZendTest\Diactoros\TestAsset\HeaderStack;
 
-class SapiEmitterTest extends TestCase
+class SapiEmitterTest extends AbstractEmitterTest
 {
-    /**
-     * @var SapiEmitter
-     */
-    protected $emitter;
-
-    public function setUp()
+    public function testEmitsBufferLevel()
     {
-        HeaderStack::reset();
-        $this->emitter = new SapiEmitter();
-    }
-
-    public function tearDown()
-    {
-        HeaderStack::reset();
-    }
-
-    public function testEmitsResponseHeaders()
-    {
+        ob_start();
+        echo "level" . ob_get_level() . " "; // 2
+        ob_start();
+        echo "level" . ob_get_level() . " "; // 3
+        ob_start();
+        echo "level" . ob_get_level() . " "; // 4
         $response = (new Response())
             ->withStatus(200)
             ->withAddedHeader('Content-Type', 'text/plain');
         $response->getBody()->write('Content!');
-
         ob_start();
         $this->emitter->emit($response);
+        $this->assertEquals('Content!', ob_get_contents());
         ob_end_clean();
-        $this->assertContains('HTTP/1.1 200 OK', HeaderStack::stack());
-        $this->assertContains('Content-Type: text/plain', HeaderStack::stack());
-        $this->assertContains('Content-Length: 8', HeaderStack::stack());
-    }
-
-    public function testEmitsMessageBody()
-    {
-        $response = (new Response())
-            ->withStatus(200)
-            ->withAddedHeader('Content-Type', 'text/plain');
-        $response->getBody()->write('Content!');
-
-        $this->expectOutputString('Content!');
-        $this->emitter->emit($response);
-    }
-
-    public function testDoesNotInjectContentLengthHeaderIfStreamSizeIsUnknown()
-    {
-        $stream = $this->prophesize('Psr\Http\Message\StreamInterface');
-        $stream->__toString()->willReturn('Content!');
-        $stream->getSize()->willReturn(null);
-        $response = (new Response())
-            ->withStatus(200)
-            ->withBody($stream->reveal());
-
-        ob_start();
-        $this->emitter->emit($response);
+        $this->assertEquals('level4 ', ob_get_contents(), 'current buffer level string must remains after emit');
         ob_end_clean();
-        foreach (HeaderStack::stack() as $header) {
-            $this->assertNotContains('Content-Length:', $header);
-        }
+        $this->emitter->emit($response, 2);
+        $this->assertEquals('level2 level3 Content!', ob_get_contents(), 'must buffer until specified level');
+        ob_end_clean();
     }
 }
